refactor: extract webhook handler
